refactored company-panel container/renderer and maintenance-panel container/renderer
