--- src/components/HostView.tsx
@@ -1,9 +1,0 @@
-import React, { useState } from "react";
-
-export const HostView = () => {
-  return (
-    <>
-      <h1>Hey this is host view</h1>
-    </>
-  );
-};
