+++ src/components/configuration-panel-renderer.tsx
@@ -1,8 +1,8 @@
-import {
-  BuildingDropdown,
-  FloorDropdown,
-} from "./components-common-utils/dynamic-buttons";
-import { DisplayConfigurableSlots } from "./ConfigureRowColSlots";
+// import {
+//   BuildingDropdown,
+//   FloorDropdown,
+// } from "./components-common-utils/dynamic-buttons";
+// import { DisplayConfigurableSlots } from "./ConfigureRowColSlots";
 
 export const ConfigureRowColSlots: React.FC = () => {
   // __inputs__:
@@ -27,11 +27,14 @@ export const ConfigureRowColSlots: React.FC = () => {
   // new floor is added:
   // set LocalPSDB by adding the new floor to that buildingIndex, floorIndex
   // --( useEffect above should auto-set totRowsIn_BF, totColsIn_BF)
-  // set confirmButtons as is
+  // set confirmButtons adding new floors confirm state as false
   // else: just shifted to existing floor
   // set totRowsIn_BF
   // set totColsIn_BF
   // set confirmButtons[]
+
+  // handleBuildingChange(buildingIndexChosen, totalBuildings):
+  // new building is added:
 
   return (
     <>
